Move action logic to actions folder
